Enter the parser...
parser.js will contain the functions necessary to parse Ableton Live projects and will also include hooks to other node.js libraries to translate note names and hopefully some method to guess which scales are used in the project. This is not yet functional and I'm pushing it merely as a conversation piece.